+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v0.4.0 (in development)
     - `AsyncJsonLinesReader`
     - `AsyncJsonLinesWriter`
     - `AsyncBufReadJsonLines` extension trait
+    - `AsyncWriteJsonLines` extension trait
 
 v0.3.0 (2022-10-30)
 -------------------